Route parsePdfToMarkdown URL sources through the SSRF guard
Extract the guard and validated-redirect fetch into src/utils/urlSafety.ts
and use it from both readFileFromUrl and loadPdfToBuffer, so the exported
parser can no longer fetch URLs the guard would refuse.
